Add files via upload
i put in my stuff--feel free to change/edit or delete anything. i still wasn't really sure how deep to go in the EDA section, so i just put a bit in there. i am also including the housing prices dataset which is from the federal reserve in order to show the graphs about the housing prices in the country.